requirements.txt
requests
beautifulsoup4
pyluach
certifi
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
PyYAML